Replace sendtestemail cmd w/ send_test_html_email.
